+++ Repetier/HAL.cpp
@@ -35,7 +35,7 @@ HAL::~HAL()
 } // ~HAL
 
 
-uint16_t HAL::integerSqrt(long a)
+uint16_t HAL::integerSqrt(int32_t a)
 {
     // http://www.mikrocontroller.net/articles/AVR_Arithmetik#32_Bit_.2F_32_Bit
     //-----------------------------------------------------------
