+++ Repetier/RF1000.h
@@ -1040,8 +1040,8 @@ Above this value the z compensation will distribute the roughness of the surface
 
 #if FEATURE_PRECISE_HEAT_BED_SCAN
 
-#define PRECISE_HEAT_BED_SCAN_WARMUP_DELAY          (unsigned long)600                                                  // [s]
-#define PRECISE_HEAT_BED_SCAN_CALIBRATION_DELAY     (unsigned long)600                                                  // [s]
+#define PRECISE_HEAT_BED_SCAN_WARMUP_DELAY          (uint32_t)600                                                  // [s]
+#define PRECISE_HEAT_BED_SCAN_CALIBRATION_DELAY     (uint32_t)600                                                  // [s]
 #define PRECISE_HEAT_BED_SCAN_BED_TEMP_PLA          60                                                                  // [°C]
 #define PRECISE_HEAT_BED_SCAN_BED_TEMP_ABS          130                                                                 // [°C]
 #define PRECISE_HEAT_BED_SCAN_EXTRUDER_TEMP_SCAN    100                                                                 // [°C]
